Implement all basic functions

+++ routes/api/boat.js
@@ -4,17 +4,38 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const loadController = require('../../controllers/loadController');
 const boatController = require('../../controllers/boatController');
+const jwtFunctions = require('../../controllers/jwtFunctions');
+// middleware jwtFunctions.checkJwt
 
 const app = express();
 app.use(bodyParser.json());
 const router = express.Router();
 
-router.post('/', function (req, res) {
-    if (req.body.hasOwnProperty('name') && req.body.hasOwnProperty('type') && req.body.hasOwnProperty('length')) {
+
+router.put('/', async function (req, res) {
+    res.status(405).json({"Error": 'PUT and DELETE on whole database is not supported'});
+});
+router.delete('/', async function (req, res) {
+    res.status(405).json({"Error": 'PUT and DELETE on whole database is not supported'});
+});
+
+router.post('/', jwtFunctions.checkJwt, function (req, res) {
+    console.log(req.user);
+
+    if (!req.user) {
+        return res.status(401).json({"Error": 'Missing or invalid jwt'});
+    }
+
+    if (req.get('content-type') !== 'application/json') {
+        res.status(415).json({"Error": 'Server only accepts application/json data'});
+    } else if (!req.accepts(['application/json'])) {
+        res.status(406).json({"Error": 'Server only responds in application/json data'});
+    } else if (req.body.hasOwnProperty('name') && req.body.hasOwnProperty('type') && req.body.hasOwnProperty('length')) {
         var name = req.body.name;
         var type = req.body.type;
         var length = req.body.length;
-        boatController.post_boat(name, type, length)
+        var owner = req.user.sub;
+        boatController.post_boat(name, type, length, owner)
             .then(key => {
                 // consider doing a get and returning a json of the actual db status
                 res.status(201).json(
@@ -24,6 +45,7 @@ router.post('/', function (req, res) {
                         "type": type,
                         "length": length,
                         "loads": [],
+                        "owner": owner,
                         "self": req.protocol + "://" + req.get("host") + req.baseUrl + "/" + key.id
                     }
                 )
@@ -33,28 +55,128 @@ router.post('/', function (req, res) {
     }
 });
 
-router.get('/:boat_id', function (req, res) {
-    boatController.get_boat(req.params.boat_id)
-        .then(boat => {
+router.get('/:boat_id', jwtFunctions.checkJwt, function (req, res) {
+    if (!req.user) {
+        return res.status(401).json({"Error": 'Missing or invalid jwt'});
+    } else if (!req.accepts(['application/json'])) {
+        res.status(406).json({"Error": 'Server only responds in application/json data'});
+    }
+    const uid = req.user.sub;
+    const id = req.params.boat_id;
+    boatController.get_boat(id)
+        .then(async boat => {
             if (boat[0] === undefined || boat[0] === null) {
-                // no boat with id
                 res.status(404).json({"Error": "No boat with this boat_id exists"});
+            } else if (boat[0].owner !== uid) {
+                res.status(403).json({"Error": "Boat is owned by different owner"});
             } else {
-                // found boat with id
                 boat[0].self = req.protocol + "://" + req.get("host") + req.baseUrl + "/" + req.params.boat_id;
                 res.status(200).json(boat[0]);
             }
         })
 });
 
-router.get('/', function (req, res) {
-    const boats = boatController.get_boats(req)
+router.get('/', jwtFunctions.checkJwt, function (req, res) {
+    // owner is null if there is no req.user
+    const owner = req.user ? req.user.sub : null;
+    if (owner === null) {
+        return res.status(401).json({"Error": 'Missing or invalid jwt'});
+    } else if (!req.accepts(['application/json'])) {
+        res.status(406).json({"Error": 'Server only responds in application/json data'});
+    }
+    const boats = boatController.get_boats(req, owner)
         .then((boats) => {
             res.status(200).json(boats);
         });
 });
 
-router.put('/:boat_id/loads/:load_id', function (req, res) {
+router.put('/:boat_id', jwtFunctions.checkJwt, function (req, res) {
+    if (!req.user) {
+        return res.status(401).json({"Error": 'Missing or invalid jwt'});
+    } else if (req.get('content-type') !== 'application/json') {
+        res.status(415).json({"Error": 'Server only accepts application/json data'});
+    } else if (!req.accepts(['application/json'])) {
+        res.status(406).json({"Error": 'Server only responds in application/json data'});
+    } else if (req.params.boat_id === "null") {
+        res.status(404).json({"Error": "No boat with this boat_id exists"});
+    } else {
+        boatController.get_boat(req.params.boat_id)
+            .then(async boat => {
+                if (boat[0] === undefined || boat[0] === null) {
+                    res.status(404).json({"Error": "No boat with this boat_id exists"});
+                } else if (boat[0].owner !== req.user.sub) {
+                    res.status(403).json({"Error": "Boat is owned by different owner"});
+                } else {
+                    if (req.body.hasOwnProperty('name') && req.body.hasOwnProperty('type') && req.body.hasOwnProperty('length')) {
+                        var name = req.body.name;
+                        var type = req.body.type;
+                        var length = req.body.length;
+                        var owner = req.user.sub;
+                        
+                        boatController.put_boat(req.params.boat_id, name, type, length, owner)
+                        .then(key => {
+                            const self = req.protocol + "://" + req.get("host") + req.baseUrl + "/" + req.params.boat_id;
+                            res.location(self);
+                            res.status(303).json(
+                                {
+                                    "id": req.params.boat_id,
+                                    "name": name,
+                                    "type": type,
+                                    "length": length,
+                                    "loads": [],
+                                    "owner": owner,
+                                    "self": req.protocol + "://" + req.get("host") + req.baseUrl + "/" + req.params.boat_id
+                                }
+                            );
+                        });
+
+                    } else {
+                        res.status(400).json({"Error": "The request object is missing at least one of the required attributes"});
+                    }
+                }
+            })
+    }
+});
+
+
+router.patch('/:boat_id', jwtFunctions.checkJwt, function (req, res) {
+    if (!req.user) {
+        return res.status(401).json({"Error": 'Missing or invalid jwt'});
+    } else if (req.get('content-type') !== 'application/json') {
+        res.status(415).json({"Error": 'Server only accepts application/json data'});
+    } else if (!req.accepts(['application/json'])) {
+        res.status(406).json({"Error": 'Server only responds in application/json data'});
+    } else if (req.params.boat_id === "null") {
+        res.status(404).json({"Error": "No boat with this boat_id exists"});
+    } else if (req.body.hasOwnProperty('id')) {
+        res.status(400).json({"Error": "boat_id cannot be patched"});
+    } else {
+        boatController.get_boat(req.params.boat_id)
+        .then(async boat => {
+            if (boat[0] === undefined || boat[0] === null) {
+                // no boat with id
+                res.status(404).json({"Error": "No boat with this boat_id exists"});
+            } else if (boat[0].owner !== req.user.sub) {
+                res.status(403).json({"Error": "Boat is owned by different owner"});
+            } else {
+                boatController.patch_boat(req.params.boat_id, req, boat)
+                .then(async key => {
+                    var self = req.protocol + "://" + req.get("host") + req.baseUrl + "/" + req.params.boat_id;
+                    res.location(self);
+                    let boat = await boatController.get_boat(req.params.boat_id);
+                    boat[0].self = self;
+                    res.status(303).json(boat[0]);
+                });
+                
+            }
+        })
+    }
+});
+
+router.put('/:boat_id/loads/:load_id', jwtFunctions.checkJwt, function (req, res) {
+    if (!req.user) {
+        return res.status(401).json({"Error": 'Missing or invalid jwt'});
+    } 
     loadController.get_load(req.params.load_id)
         .then(load => {
             if (load[0] === undefined || load[0] === null) {
@@ -67,6 +189,8 @@ router.put('/:boat_id/loads/:load_id', function (req, res) {
                         if (boat[0] === undefined || boat[0] === null) {
                             // no boat with id
                             res.status(404).json({"Error": "The specified boat and/or load does not exist"});
+                        } else if (boat[0].owner !== req.user.sub) {
+                            res.status(403).json({"Error": "Boat is owned by different owner"});
                         } else {
                             // found the boat with id
                             // load unassigned
@@ -82,7 +206,10 @@ router.put('/:boat_id/loads/:load_id', function (req, res) {
         })
 });
 
-router.delete('/:boat_id/loads/:load_id', function (req, res) {
+router.delete('/:boat_id/loads/:load_id', jwtFunctions.checkJwt, function (req, res) {
+    if (!req.user) {
+        return res.status(401).json({"Error": 'Missing or invalid jwt'});
+    }
     loadController.get_load(req.params.load_id)
         .then(load => {
             if (load[0] === undefined || load[0] === null) {
@@ -95,6 +222,8 @@ router.delete('/:boat_id/loads/:load_id', function (req, res) {
                         if (boat[0] === undefined || boat[0] === null) {
                             // no boat with id
                             res.status(404).json({"Error": "No boat with this boat_id is loaded with the load with this load_id"});
+                        } else if (boat[0].owner !== req.user.sub) {
+                            res.status(403).json({"Error": "Boat is owned by different owner"});
                         } else {
                             // found the boat with id
                             var found = false;
@@ -115,12 +244,17 @@ router.delete('/:boat_id/loads/:load_id', function (req, res) {
         })
 });
 
-router.delete('/:boat_id', function (req, res) {
+router.delete('/:boat_id', jwtFunctions.checkJwt, function (req, res) {
+    if (!req.user) {
+        return res.status(401).json({"Error": 'Missing or invalid jwt'});
+    }
     boatController.get_boat(req.params.boat_id)
         .then(async boat => {
             if (boat[0] === undefined || boat[0] === null) {
                 // no boat with id
                 res.status(404).json({"Error": "No boat with this boat_id exists"});
+            } else if (boat[0].owner !== req.user.sub) {
+                res.status(403).json({"Error": "Boat is owned by different owner"});
             } else {
                 // found boat with id
                 await boatController.delete_boat(req.params.boat_id)
@@ -129,12 +263,19 @@ router.delete('/:boat_id', function (req, res) {
         })
 });
 
-router.get('/:boat_id/loads', function (req, res) {
+router.get('/:boat_id/loads', jwtFunctions.checkJwt, function (req, res) {
+    if (!req.user) {
+        return res.status(401).json({"Error": 'Missing or invalid jwt'});
+    } else if (!req.accepts(['application/json'])) {
+        res.status(406).json({"Error": 'Server only responds in application/json data'});
+    }
     boatController.get_boat(req.params.boat_id)
         .then(boat => {
             if (boat[0] === undefined || boat[0] === null) {
                 // no boat with id
                 res.status(404).json({"Error": "No boat with this boat_id exists"});
+            } else if (boat[0].owner !== req.user.sub) {
+                res.status(403).json({"Error": "Boat is owned by different owner"});
             } else {
                 // found boat with id
                 boatController.get_boat_loads(req, req.params.boat_id)
